+++ src/recipes/mod.rs
@@ -1,3 +1,4 @@
 //! Extended ZooKeeper recipes from [Apache Curator](http://curator.apache.org/).
 pub mod cache;
+pub mod leader_latch;
 pub mod shared_lock;